Use polyglossia instead of babel when compiling with Xe(La)TeX
XeTeX and XeLaTeX use polyglossia instead of babel, so use that instead
when compiling with either of those two, and otherwise use babel like
normal.

This seems to change some other behaviour as well: the font of the main
body is now BookmanOldStyle instead of Arial. By looking at the code,
this seems to be the intended behaviour (see the line with
the "\setmainfont" command).

I removed the driver options ([xetex]) as per this answer on
TeX.stackexchange: http://tex.stackexchange.com/a/2985.

In the case of the graphicx package, the xetex option was removed
because it produced an error for me due to conflicting package options:
[xetex] vs [] (i.e. no options). This error only arose after switching
to polyglossia. This leads me to believe that indeed it is not used in
practice (since another package decided to load it without the xetex
driver option).